requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
anyio==4.13.0
asgiref==3.11.1
astunparse==1.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
Django==4.2.7
django-dotenv==1.4.2
exceptiongroup==1.3.1
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.3.0
gast==0.7.0
google-auth==2.49.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.80.0
gunicorn==21.2.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
keras==2.14.0
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
namex==0.1.0
networkx==3.4.2
numpy==1.24.4
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
Pillow==10.1.0
protobuf==4.25.9
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.20.0
python-decouple==3.8
PyYAML==6.0.3
regex==2026.4.4
requests==2.31.0
requests-oauthlib==2.0.0
rich==15.0.0
safetensors==0.7.0
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.5
sympy==1.14.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.14.0
termcolor==3.3.0
tokenizers==0.15.2
torch==2.1.0
tqdm==4.67.3
transformers==4.35.2
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
whitenoise==6.6.0
wrapt==1.14.2
